[fixed] Make the non-minified dist build present again (#164)
This commit removes the `-p` flag from the call to webpack
so that the webpack config works as intended, creating two
files in the dist directory, one minified and one unminified.

After 3f323381b1e8c43fdbf172813b9e8a4ace9abe79 was merged the
build ended up only producing the production minified versions
for both files.  This commit restores the initial behavior
of a non-minified and a minified file.